skog-ai/requirements.txt
openai
faiss-cpu
streamlit
PyPDF2
python-dotenv
numpy
